KEYCLOAK-16077 Remove need for MapStorage.replace
