+++ model/map/src/main/java/org/keycloak/models/map/client/MapClientProvider.java
@@ -19,12 +19,12 @@ package org.keycloak.models.map.client;
 
 import org.jboss.logging.Logger;
 import org.keycloak.models.ClientModel;
+import org.keycloak.models.ClientModel.ClientUpdatedEvent;
 import org.keycloak.models.ClientProvider;
 import org.keycloak.models.KeycloakSession;
 import org.keycloak.models.ModelDuplicateException;
 import org.keycloak.models.RealmModel;
 
-import org.keycloak.models.RealmModel.ClientUpdatedEvent;
 import org.keycloak.models.map.storage.MapKeycloakTransaction;
 import org.keycloak.models.map.common.Serialization;
 import java.util.Comparator;
@@ -71,7 +71,7 @@ public class MapClientProvider implements ClientProvider {
     }
 
     private ClientUpdatedEvent clientUpdatedEvent(ClientModel c) {
-        return new RealmModel.ClientUpdatedEvent() {
+        return new ClientModel.ClientUpdatedEvent() {
             @Override
             public ClientModel getUpdatedClient() {
                 return c;
@@ -96,8 +96,6 @@ public class MapClientProvider implements ClientProvider {
         return origEntity -> new MapClientAdapter(session, realm, registerEntityForChanges(origEntity)) {
             @Override
             public void updateClient() {
-                // commit
-                MapClientProvider.this.tx.replace(entity.getId(), this.entity);
                 session.getKeycloakSessionFactory().publish(clientUpdatedEvent(this));
             }
 
@@ -178,7 +176,7 @@ public class MapClientProvider implements ClientProvider {
         final ClientModel resource = entityToAdapterFunc(realm).apply(entity);
 
         // TODO: Sending an event should be extracted to store layer
-        session.getKeycloakSessionFactory().publish((RealmModel.ClientCreationEvent) () -> resource);
+        session.getKeycloakSessionFactory().publish((ClientModel.ClientCreationEvent) () -> resource);
         resource.updateClient();        // This is actualy strange contract - it should be the store code to call updateClient
 
         return resource;
@@ -214,7 +212,7 @@ public class MapClientProvider implements ClientProvider {
         session.users().preRemove(realm, client);
         session.roles().removeRoles(client);
 
-        session.getKeycloakSessionFactory().publish(new RealmModel.ClientRemovedEvent() {
+        session.getKeycloakSessionFactory().publish(new ClientModel.ClientRemovedEvent() {
             @Override
             public ClientModel getClient() {
                 return client;
