+++ model/map/src/main/java/org/keycloak/models/map/storage/MapStorage.java
@@ -34,8 +34,6 @@ public interface MapStorage<K, V> {
 
     V remove(K key);
 
-    V replace(K key, V value);
-
     Set<K> keySet();
 
     Set<Map.Entry<K,V>> entrySet();
